# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Social media platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: string;
  username: string;
  display_name: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatar_url?: string | null;
  /** @nullable */
  website?: string | null;
  created_at: string;
  followers_count: number;
  following_count: number;
  posts_count: number;
  is_following?: boolean;
}

export interface UserSummary {
  id: string;
  username: string;
  display_name: string;
  /** @nullable */
  avatar_url?: string | null;
  /** @nullable */
  bio?: string | null;
  is_following?: boolean;
  followers_count?: number;
}

export interface ProfileUpdate {
  display_name?: string;
  bio?: string;
  avatar_url?: string;
  website?: string;
}

export interface Post {
  id: string;
  content: string;
  /** @nullable */
  image_url?: string | null;
  author: UserSummary;
  created_at: string;
  /** @nullable */
  updated_at?: string | null;
  likes_count: number;
  comments_count: number;
  is_liked?: boolean;
}

export interface PostInput {
  /**
     * @minLength 1
     * @maxLength 2200
     */
  content: string;
  image_url?: string;
}

export interface PostUpdate {
  /**
     * @minLength 1
     * @maxLength 2200
     */
  content?: string;
}

export interface PostsPage {
  posts: Post[];
  /** @nullable */
  next_cursor?: string | null;
  has_more: boolean;
}

export interface Comment {
  id: string;
  content: string;
  post_id: string;
  author: UserSummary;
  created_at: string;
}

export interface CommentInput {
  /**
     * @minLength 1
     * @maxLength 1000
     */
  content: string;
}

export interface LikeResult {
  liked: boolean;
  likes_count: number;
}

export interface FollowResult {
  following: boolean;
  followers_count: number;
}

export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export const NotificationType = {
  like: 'like',
  comment: 'comment',
  follow: 'follow',
  mention: 'mention',
} as const;

export interface Notification {
  id: string;
  type: NotificationType;
  actor?: UserSummary;
  /** @nullable */
  post_id?: string | null;
  /** @nullable */
  post_content?: string | null;
  /** @nullable */
  comment_content?: string | null;
  created_at: string;
  is_read: boolean;
}

export interface NotificationsPage {
  notifications: Notification[];
  /** @nullable */
  next_cursor?: string | null;
  has_more: boolean;
}

export interface MarkReadResult {
  updated: number;
}

export interface UnreadCount {
  count: number;
}

export type FindUsersParams = {
q: string;
};