# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Social media platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}/posts:
    get:
      operationId: listUserPosts
      tags: [users]
      summary: Get posts by user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsPage"
...
  /posts:
    get:
      operationId: listFeed
      tags: [feed]
      summary: Get global feed
      responses:
        "200":
          description: Feed posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsPage"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{postId}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a post by ID
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /posts/{postId}/like:
    post:
      operationId: likePost
      tags: [likes]
      summary: Like a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Like result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
    delete:
      operationId: unlikePost
      tags: [likes]
      summary: Unlike a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unlike result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
  /posts/{postId}/comments:
    get:
      operationId: getComments
      tags: [comments]
      summary: Get comments on a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Create a comment
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /follows/{targetUserId}:
    post:
      operationId: followUser
      tags: [follows]
      summary: Follow a user
      parameters:
        - name: targetUserId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Follow result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResult"
    delete:
      operationId: unfollowUser
      tags: [follows]
      summary: Unfollow a user
      parameters:
        - name: targetUserId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unfollow result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResult"
...
components:
  schemas:
...
    UserSummary:
      type: object
      required: [id, username, display_name]
      properties:
        id:
          type: string
        username:
          type: string
        display_name:
          type: string
        avatar_url:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        is_following:
          type: boolean
        followers_count:
          type: integer
...
    Post:
      type: object
      required: [id, content, author, created_at, likes_count, comments_count]
      properties:
        id:
          type: string
        content:
          type: string
        image_url:
          type: ["string", "null"]
        author:
          $ref: "#/components/schemas/UserSummary"
        created_at:
          type: string
        updated_at:
          type: ["string", "null"]
        likes_count:
          type: integer
        comments_count:
          type: integer
        is_liked:
          type: boolean
    PostInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 2200
        image_url:
          type: string
    PostUpdate:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 2200
    PostsPage:
      type: object
      required: [posts, has_more]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        next_cursor:
          type: ["string", "null"]
        has_more:
          type: boolean
    Comment:
      type: object
      required: [id, content, author, post_id, created_at]
      properties:
        id:
          type: string
        content:
          type: string
        post_id:
          type: string
        author:
          $ref: "#/components/schemas/UserSummary"
        created_at:
          type: string
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000
    LikeResult:
      type: object
      required: [liked, likes_count]
      properties:
        liked:
          type: boolean
        likes_count:
          type: integer
    FollowResult:
      type: object
      required: [following, followers_count]
      properties:
        following:
          type: boolean
        followers_count:
          type: integer
...