# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Social media platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserSummary } from './userSummary';

export interface Comment {
  id: string;
  content: string;
  post_id: string;
  author: UserSummary;
  created_at: string;
}