# lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Social media platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PostInput {
  /**
     * @minLength 1
     * @maxLength 2200
     */
  content: string;
  image_url?: string;
}