# lib/api-zod/src/generated/types/postUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Social media platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PostUpdate {
  /**
     * @minLength 1
     * @maxLength 2200
     */
  content?: string;
}