# lib/api-zod/src/generated/types/userSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Social media platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UserSummary {
  id: string;
  username: string;
  display_name: string;
  /** @nullable */
  avatar_url?: string | null;
  /** @nullable */
  bio?: string | null;
  is_following?: boolean;
  followers_count?: number;
}